+++ homework_05/.dockerignore
@@ -1,0 +1,7 @@
+__pycache__
+.venv
+test
+Dockerfile
+myapp.log
+.dockerignore
+README.md
