+++ homework_05/Dockerfile
@@ -3,10 +3,10 @@ FROM python:3.11-slim
 WORKDIR /app
 
 
-COPY ./requirements.txt ./requirements.txt
+COPY requirements.txt .
 
 
-RUN pip install --no-cashe-dir requirements.txt
+RUN  pip install --no-cache-dir -r requirements.txt
 
 
 COPY . .
