+++ homework_05/README.md
@@ -1,0 +1,21 @@
+Приложение FastAPI в контейнере
+Приложение представляет список книг для последующего прочтения,
+это список из 6 книг.
+Возможно добавдение книг через API
+При перезапуске приложения, список сбрасываентся к стандартому, заренее подготовленному
+
+Создать контейнер
+docker build -t hw05 .
+
+Запустить контейнер
+docker start hw05
+
+Приложение отвечает по всем адресам по птротоколу http на порту TCP 8000
+Основной интефейс 
+127.0.0.1:8000
+
+API
+127.0.0.1:8000/api
+127.0.0.1:8000/books/1
+127.0.0.1:8000/ping
+
