+++ homework_05/__init__.py
@@ -13,7 +13,7 @@ __all__ = [
 def main():
     logging.basicConfig(filename='myapp.log', level=logging.INFO)
     logger.info('Started')
-    uvicorn.run("app:app", host='127.0.0.1', port=8000)
+    uvicorn.run("app:app", host='0.0.0.0', port=8000)
     logger.info('Finished')
